+++ apps/backend/src/modules/projects/project.controller.ts
@@ -78,7 +78,7 @@ class ProjectController extends BaseController {
 
 		this.addRoute({
 			handler: (options) =>
-				this.findAllByName(
+				this.findAll(
 					options as APIHandlerOptions<{
 						query: ProjectGetAllRequestDto;
 					}>,
@@ -189,7 +189,13 @@ class ProjectController extends BaseController {
 	 * @swagger
 	 * /projects:
 	 *    get:
-	 *      description: Returns an array of projects
+	 *      description: Returns all projects or filters them by name if the 'name' query parameter is provided.
+	 *      parameters:
+	 *        - in: query
+	 *          name: name
+	 *          schema:
+	 *            type: string
+	 *          description: The name to filter projects by (optional).
 	 *      responses:
 	 *        200:
 	 *          description: Successful operation
@@ -203,15 +209,15 @@ class ProjectController extends BaseController {
 	 *                    items:
 	 *                      $ref: "#/components/schemas/Project"
 	 */
-	private async findAllByName(
+	private async findAll(
 		options: APIHandlerOptions<{
 			query: ProjectGetAllRequestDto;
 		}>,
 	): Promise<APIHandlerResponse> {
-		const { query } = options;
+		const { name } = options.query;
 
 		return {
-			payload: await this.projectService.findAllbyName(query),
+			payload: await this.projectService.findAll(name),
 			status: HTTPCode.OK,
 		};
 	}