+++ apps/backend/src/modules/projects/project.service.ts
@@ -7,7 +7,6 @@ import { ProjectError } from "./libs/exceptions/exceptions.js";
 import {
 	type ProjectCreateRequestDto,
 	type ProjectGetAllItemResponseDto,
-	type ProjectGetAllRequestDto,
 	type ProjectGetAllResponseDto,
 	type ProjectPatchRequestDto,
 	type ProjectPatchResponseDto,
@@ -91,21 +90,11 @@ class ProjectService implements Service {
 		return await this.injectProjectApiKeyToProject(item.toObject());
 	}
 
-	public async findAll(): Promise<ProjectGetAllResponseDto> {
-		const projects = await this.projectRepository.findAll();
+	public async findAll(name?: string): Promise<ProjectGetAllResponseDto> {
+		const projects = await this.projectRepository.findAll(name);
 
 		return {
 			items: projects.items.map((item) => item.toObject()),
-		};
-	}
-
-	public async findAllbyName(
-		query: ProjectGetAllRequestDto,
-	): Promise<ProjectGetAllResponseDto> {
-		const items = await this.projectRepository.findAllbyName(query.name);
-
-		return {
-			items: items.map((item) => item.toObject()),
 		};
 	}