+++ apps/frontend/src/libs/enums/notification-message.enum.ts
@@ -1,7 +1,8 @@
 const NotificationMessage = {
+	GROUP_CREATE_SUCCESS: "Group was successfully created.",
 	PROFILE_UPDATE_SUCCESS: "Successfully updated profile information.",
 	PROJECT_API_KEY_CREATE_SUCCESS: "Project API key was successfully generated.",
-	PROJECT_CREATE_SUCCESS: "Project was successfully created",
+	PROJECT_CREATE_SUCCESS: "Project was successfully created.",
 	PROJECT_DELETE_SUCCESS: "Project was successfully deleted.",
 	PROJECT_UPDATE_SUCCESS: "Project was successfully updated.",
 	USER_DELETE_SUCCESS: "User deleted successfully.",