+++ apps/frontend/src/libs/hooks/hooks.ts
@@ -5,6 +5,8 @@ export { useHandleClickOutside } from "./use-handle-click-outside/use-handle-cli
 export { useModal } from "./use-modal-state/use-modal-state.hook.js";
 export { usePagination } from "./use-pagination/use-pagination.hook.js";
 export { usePopover } from "./use-popover-state/use-popover-state.hook.js";
+export { useSearch } from "./use-search/use-search.hook.js";
+export { useSelectedItems } from "./use-selected-items/use-selected-items.hook.js";
 export { useCallback, useEffect, useMemo, useRef, useState } from "react";
 export {
 	useController as useFormController,