+++ apps/frontend/src/modules/groups/libs/types/types.ts
@@ -1,4 +1,6 @@
 export {
+	type GroupCreateRequestDto,
+	type GroupCreateResponseDto,
 	type GroupGetAllItemResponseDto,
 	type GroupGetAllResponseDto,
 } from "@git-fit/shared";