+++ packages/shared/src/index.ts
@@ -79,6 +79,7 @@ export {
 	type ProjectPatchResponseDto,
 	projectPatchValidationSchema,
 	ProjectsApiPath,
+	ProjectValidationRule,
 } from "./modules/projects/projects.js";
 export {
 	type UserAuthResponseDto,